Application Updated Now User will be accessed with JWT Token

+++ src/main/java/com/emp/application/controller/Controller.java
@@ -6,18 +6,21 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.emp.application.dao.Dto;
 import com.emp.application.entity.User;
 import com.emp.application.service.Service;
 
+import jakarta.validation.Valid;
+
 @RequestMapping("/dashboard")
 @RestController
-@CrossOrigin("http://localhost:5173")
 public class Controller {
 	
 	@Autowired
@@ -28,10 +31,17 @@ public class Controller {
 		return principal.getName();		
 	}
 	
-	@PostMapping("/delete-user")
-	public String deleteUserController(long id) {
-		return service.deleteUserService(id);
+	@DeleteMapping("/delete-user")
+	public ResponseEntity<?> deleteUserController(@RequestBody Dto request) {
+	    Long id = request.getId();
+	    if (id != null && id != 0) {
+	        System.out.println("Deleting user with ID: " + id);
+	        service.deleteUserService(id);
+	        return ResponseEntity.ok().build();
+	    }
+	    return ResponseEntity.badRequest().build();
 	}
+
 	
 	@GetMapping("/get-users")
 	public ResponseEntity<List<User>> getUsersController() {
@@ -45,6 +55,15 @@ public class Controller {
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		}
 	}
+	
+	@PostMapping("/update-user")
+	public ResponseEntity<?> updateUserController(@Valid @RequestBody Dto dto) {
+		System.out.println("DTO received: " + dto); // log actual contents
+	    System.out.println("Updating user: " + dto.getEmail());
+	    service.updateUserService(dto);
+	    return ResponseEntity.ok("User updated successfully");
+	}
+
 	
 	
 }
